TFP-4873 rydde litt kode før yf-rydding
